Refactor : 퍼렐럴 라우팅 수정
