+++ src/app/(content-layout)/[groupId]/tasklist/@detailTask/default.tsx
@@ -1,0 +1,3 @@
+export default function DefaultRender() {
+  return null;
+}
