+++ src/app/(content-layout)/[groupId]/tasklist/_tasklist/components/TasksWiseTask.tsx
@@ -11,8 +11,7 @@ import getDetailTaskItem from '@/lib/api/detail-task-item';
 import { useSortable } from '@dnd-kit/sortable';
 import { CSS } from '@dnd-kit/utilities';
 import { useDndMonitor } from '@dnd-kit/core';
-import Link from 'next/link';
-import DetailTaskContainer from '../../@detailTask/(..)tasks/[taskId]/page';
+import { useRouter } from 'next/navigation';
 
 interface Props {
   task: Task;
@@ -23,13 +22,14 @@ interface Props {
 export default function TasksWiseTask({ task, groupId, taskListId }: Props) {
   const [isDone, setIsDone] = useState(!!task.doneAt);
   const [isDelete, setIsDelete] = useState(false);
-  const [isDetailTaskOpen, setIsDetailTaskOpen] = useState(false);
   const [detailTask, setDetailTask] = useState<DetailTaskType>();
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
   const { attributes, listeners, setNodeRef, transform, transition, isDragging } = useSortable({
     id: task.id,
   });
   const [onDrag, setOnDrag] = useState(false);
+
+  const router = useRouter();
 
   useDndMonitor({
     onDragStart: () => setOnDrag(true),
@@ -78,11 +78,7 @@ export default function TasksWiseTask({ task, groupId, taskListId }: Props) {
   };
 
   const openDetailTask = () => {
-    setIsDetailTaskOpen(true);
-  };
-
-  const closeDetailTask = () => {
-    setIsDetailTaskOpen(false);
+    router.push(`/${groupId}/tasks/${task.id}`);
   };
 
   return (
@@ -98,28 +94,19 @@ export default function TasksWiseTask({ task, groupId, taskListId }: Props) {
             opacity: isDragging ? 0.5 : 1,
           }}
         >
-          <Link href={`/${groupId}/tasks/${task.id}`}>
-            <TaskListItem
-              key={task.id}
-              type="taskList"
-              checkDropdownOpen={checkDropdownOpen}
-              onCheckStatusChange={() => toggleTaskDone(isDone, toggleTaskStatus)}
-              onEdit={() => popUpEditTaskModal(createOrEditModalId)}
-              onDelete={() => popUpDeleteTaskModal(taskDeleteModalId)}
-              onClick={() => openDetailTask()}
-              isDone={isDone}
-              name={task.name}
-              commentCount={task.commentCount}
-              date={safeFormatDate(task.date)}
-              frequency={task.frequency}
-            />
-          </Link>
-          <DetailTaskContainer
+          <TaskListItem
+            key={task.id}
+            type="taskList"
+            checkDropdownOpen={checkDropdownOpen}
+            onCheckStatusChange={() => toggleTaskDone(isDone, toggleTaskStatus)}
+            onEdit={() => popUpEditTaskModal(createOrEditModalId)}
+            onDelete={() => popUpDeleteTaskModal(taskDeleteModalId)}
             isDone={isDone}
-            setIsDone={toggleTaskStatus}
-            taskId={task.id}
-            closeDetailTask={closeDetailTask}
-            isOpen={isDetailTaskOpen}
+            onClick={() => openDetailTask()}
+            name={task.name}
+            commentCount={task.commentCount}
+            date={safeFormatDate(task.date)}
+            frequency={task.frequency}
           />
           <RemoveTaskModal
             taskName={task.name}
