+++ src/app/(content-layout)/[groupId]/tasklist/_tasklist/hooks/use-task-actions.ts
@@ -1,7 +1,7 @@
 import axiosClient from '@/lib/axiosClient';
 import { Task } from '../types/task-type';
 import { Toast } from '@/components/common/Toastify';
-import { revalidateDetailTask } from '../actions/task-actions';
+import { revalidateDetailTask, revalidateTasks } from '../actions/task-actions';
 
 export function useTaskActions(task?: Task) {
   const deleteTask = async (
