+++ src/app/(content-layout)/[groupId]/tasklist/layout.tsx
@@ -1,10 +1,10 @@
 export default function TaskListPageLayout({
   children,
   detailTask,
-}: Readonly<{
+}: {
   children: React.ReactNode;
   detailTask: React.ReactNode;
-}>) {
+}) {
   return (
     <>
       {children}
