+++ src/app/(content-layout)/[groupId]/tasks/[taskId]/page.tsx
@@ -1,3 +1,4 @@
+import { ErrorBoundary } from 'react-error-boundary';
 import DetailTaskCommentField from '../../tasklist/@detailTask/_components/DetailTaskCommentsField';
 import Content from '../../tasklist/@detailTask/_components/DetailTaskContentField';
 import ToggleDoneButton from '../../tasklist/@detailTask/_components/ToggleDoneButton';
@@ -15,10 +16,12 @@ export default async function DetailTaskPage({ params }: Props) {
   const isDone = Boolean(task.doneAt);
 
   return (
-    <div className="flex h-full w-full flex-col gap-25 overflow-scroll">
-      <Content isDone={isDone} task={task} />
-      <DetailTaskCommentField taskId={task.id} />
-      <ToggleDoneButton isDone={isDone} task={task} />
-    </div>
+    <ErrorBoundary fallback={<div>해당 태스크를 불러올 수 없습니다.</div>}>
+      <div className="flex h-full w-full flex-col gap-25 overflow-scroll">
+        <Content isDone={isDone} task={task} />
+        <DetailTaskCommentField taskId={task.id} />
+        <ToggleDoneButton isDone={isDone} task={task} />
+      </div>
+    </ErrorBoundary>
   );
 }
